GrapesJS installed, selected important plugins
