+++ client/package.json
@@ -15,6 +15,7 @@
     "@testing-library/user-event": "^13.5.0",
     "axios": "^1.4.0",
     "bootstrap": "^5.3.0",
+    "file-saver": "^2.0.5",
     "grapesjs": "^0.21.3",
     "grapesjs-blocks-basic": "^1.0.2",
     "grapesjs-blocks-bootstrap4": "^0.2.3",
@@ -32,7 +33,9 @@
     "grapesjs-tabs": "^1.0.6",
     "grapesjs-tooltip": "^0.1.8",
     "grapesjs-tui-image-editor": "^1.0.2",
+    "jszip": "^3.10.1",
     "jwt-decode": "^3.1.2",
+    "notyf": "^3.10.0",
     "react": "^18.2.0",
     "react-bootstrap": "^2.7.4",
     "react-dom": "^18.2.0",
