+++ client/src/App.js
@@ -5,7 +5,7 @@ import Home from "./pagecomponents/HomePage.js";
 import RegisterForm from "./pagecomponents/RegisterForm.js";
 import LoginForm from "./pagecomponents/LoginForm";
 import About from "./pagecomponents/AboutPage";
-import Editor from "./pagecomponents/EditorPage";
+import Editor from "./pagecomponents/editorpage/EditorPage";
 import RequireAuth from "./pagecomponents/authcontrollers/RequireAuth";
 import PurchasePlan from "./pagecomponents/PurchasePlan";
 import Deploy from "./pagecomponents/DeployPage";
@@ -20,9 +20,10 @@ function App() {
         <Route path="/register" element={<RegisterForm />} />
         <Route path="/purchaseplan" element={<PurchasePlan />} />
         <Route path="/login" element={<LoginForm />} />
+        <Route path="/editor" element={<Editor />} />
+
         <Route element={<PersistLogin />}>
           <Route element={<RequireAuth allowedRoles={[1974]} />}>
-            <Route path="/editor" element={<Editor />} />
             <Route path="/about" element={<About />} />{" "}
           </Route>
           <Route element={<RequireAuth allowedRoles={[5150]} />}>
