+++ client/src/pagecomponents/editorpage/grapejsAssets/buttonBlock.js
@@ -1,0 +1,17 @@
+export const buttonBlockConfig = {
+  id: "button",
+  label: "Button",
+  attributes: { class: "fas fa-hand-pointer fa-3x" },
+  category: "Content",
+  content: {
+    type: "link",
+    content: "Button",
+    style: {
+      display: "inline-block",
+      padding: "10px",
+      textDecoration: "none",
+      background: "#3f51b5",
+      color: "#ffffff",
+    },
+  },
+};
