+++ client/src/pagecomponents/editorpage/grapejsAssets/imageBlock.js
@@ -1,0 +1,9 @@
+export const imageBlockConfig = {
+  id: "image",
+  label: "Image",
+  category: "Content",
+
+  attributes: { class: "fas fa-camera fa-3x" },
+  select: true,
+  content: { type: "image" },
+};
