+++ client/src/pagecomponents/editorpage/grapejsAssets/textBlock.js
@@ -1,0 +1,8 @@
+export const textBlockConfig = {
+  id: "text",
+  label: "Text",
+  category: "Content",
+
+  attributes: { class: "fas fa-font fa-3x" },
+  content: '<div data-gjs-type="text">Insert your text here</div>',
+};
